Refactor option conversions.
