+++ src/main/scala/com/mwz/sonar/scala/util/PathUtils.scala
@@ -20,32 +20,10 @@ package com.mwz.sonar.scala
 package util
 
 import java.nio.file.{Path, Paths}
-import java.util.Optional
 
 import org.sonar.api.batch.fs.FileSystem
 
 import scala.language.implicitConversions
-
-/**
- *  Scala.Option <-> Java.Optional conversions.
- *  @note Taken from https://gist.github.com/julienroubieu/fbb7e1467ab44203a09f.
- */
-object JavaOptionals {
-  implicit final def toRichOption[T >: Null](opt: Option[T]): RichOption[T] =
-    new RichOption[T](opt)
-  implicit final def toRichOptional[T](optional: Optional[T]): RichOptional[T] =
-    new RichOptional[T](optional)
-}
-
-/** Transform this Option to an equivalent Java Optional */
-final class RichOption[T >: Null](val opt: Option[T]) extends AnyVal {
-  def toOptional: Optional[T] = Optional.ofNullable(opt.orNull)
-}
-
-/** Transform this Optional to an equivalent Scala Option */
-final class RichOptional[T](val opt: Optional[T]) extends AnyVal {
-  def toOption: Option[T] = if (opt.isPresent) Some(opt.get()) else None
-}
 
 /**
  * Various Path utilities.
