+++ src/main/scala/com/mwz/sonar/scala/util/syntax/SonarFileSystem.scala
@@ -20,26 +20,28 @@ package com.mwz.sonar.scala
 package util
 package syntax
 
-import org.sonar.api.batch.fs.InputFile
-import org.sonar.api.batch.measure.Metric
-import org.sonar.api.batch.sensor.SensorContext
+import java.io.File
+import java.nio.file.Path
 
-object SensorContext {
-  implicit final class SensorContextOps(val context: SensorContext) extends AnyVal {
+import cats.syntax.flatMap._
+import cats.{Monad, MonoidK}
+import org.sonar.api.batch.fs.FileSystem
+
+import scala.language.higherKinds
+import scala.util.{Failure, Success, Try}
+
+object SonarFileSystem {
+  implicit final class FileSystemOps(val fs: FileSystem) extends AnyVal {
 
     /**
-     * Save a new measure for the given metric.
+     * Resolve paths relative to the given file system.
      */
-    def saveMeasure[T <: java.io.Serializable](
-      file: InputFile,
-      metric: Metric[T],
-      value: T
-    ): Unit =
-      context
-        .newMeasure[T]
-        .on(file)
-        .forMetric(metric)
-        .withValue(value)
-        .save()
+    def resolve[F[_]: Monad: MonoidK](toResolve: F[Path]): F[File] =
+      toResolve.flatMap[File] { path =>
+        Try(fs.resolvePath(path.toString)) match {
+          case Failure(_) => MonoidK[F].empty
+          case Success(f) => Monad[F].pure(f)
+        }
+      }
   }
 }
