+++ src/main/scala/com/mwz/sonar/scala/util/syntax/SonarSensorContext.scala
@@ -20,26 +20,26 @@ package com.mwz.sonar.scala
 package util
 package syntax
 
-import java.nio.file.{Path, Paths}
+import org.sonar.api.batch.fs.InputFile
+import org.sonar.api.batch.measure.Metric
+import org.sonar.api.batch.sensor.SensorContext
 
-import cats.instances.list._
-import cats.instances.option._
-import com.mwz.sonar.scala.util.syntax.FileSystem._
-import org.scalatest.{FlatSpec, Matchers, OptionValues}
-import org.sonar.api.batch.fs.internal.DefaultFileSystem
+object SonarSensorContext {
+  implicit final class SensorContextOps(val context: SensorContext) extends AnyVal {
 
-class FileSystemSpec extends FlatSpec with Matchers with OptionValues {
-  it should "attempt to resolve paths" in {
-    val fs = new DefaultFileSystem(Paths.get("./"))
-
-    val paths = List(Paths.get("path/1"), Paths.get("path/2"))
-    fs.resolve(paths) shouldBe List(
-      Paths.get("./").resolve("path/1").toAbsolutePath.normalize.toFile,
-      Paths.get("./").resolve("path/2").toAbsolutePath.normalize.toFile
-    )
-
-    val path: Option[Path] = Some(Paths.get("another/path"))
-    fs.resolve(path).value shouldBe
-    Paths.get("./").resolve("another/path").toAbsolutePath.normalize.toFile
+    /**
+     * Save a new measure for the given metric.
+     */
+    def saveMeasure[T <: java.io.Serializable](
+      file: InputFile,
+      metric: Metric[T],
+      value: T
+    ): Unit =
+      context
+        .newMeasure[T]
+        .on(file)
+        .forMetric(metric)
+        .withValue(value)
+        .save()
   }
 }
