+++ src/test/scala/com/mwz/sonar/scala/util/syntax/SonarSensorContextSpec.scala
@@ -22,13 +22,13 @@ package syntax
 
 import java.nio.file.Paths
 
-import com.mwz.sonar.scala.util.syntax.SensorContext._
+import com.mwz.sonar.scala.util.syntax.SonarSensorContext._
 import org.scalatest.{FlatSpec, Matchers}
 import org.sonar.api.batch.fs.internal.TestInputFileBuilder
 import org.sonar.api.batch.sensor.internal.SensorContextTester
 import org.sonar.api.measures.CoreMetrics
 
-class SensorContextSpec extends FlatSpec with Matchers with SensorContextMatchers {
+class SonarSensorContextSpec extends FlatSpec with Matchers with SensorContextMatchers {
   it should "save a measure for a given input file" in {
     val ctx = SensorContextTester.create(Paths.get("./"))
     val testFile = TestInputFileBuilder
